+++ lib/product/widget/card/news_card.dart
@@ -3,100 +3,43 @@ import 'package:kartal/kartal.dart';
 import 'package:life_shared/life_shared.dart';
 import 'package:vbaseproject/product/common/color_common.dart';
 import 'package:vbaseproject/product/package/custom_network_image.dart';
-import 'package:vbaseproject/product/utility/decorations/empty_box.dart';
-import 'package:vbaseproject/product/utility/formatter/custom_date_time_formatter.dart';
+import 'package:vbaseproject/product/utility/decorations/custom_radius.dart';
 import 'package:vbaseproject/product/utility/padding/page_padding.dart';
 
-class NewsCard extends StatelessWidget {
-  const NewsCard({
-    required this.item,
-    required this.onTap,
-    super.key,
-  });
+final class NewsCard extends StatelessWidget {
+  const NewsCard({required this.item, required this.onTap, super.key});
+
   final NewsModel item;
   final VoidCallback onTap;
 
   @override
   Widget build(BuildContext context) {
-    return InkWell(
-      onTap: onTap,
-      child: SizedBox(
-        height: context.sized.dynamicHeight(.45),
-        child: Stack(
-          children: [
-            Positioned.fill(
-              child: _NewsImage(item: item),
-            ),
-            Positioned(
-              bottom: 0,
-              left: 0,
-              right: 0,
-              child: _GradientBox(item: item),
-            ),
-            Positioned(
-              bottom: 0,
-              left: 0,
-              right: 0,
-              child: _TextArea(item: item),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _GradientBox extends StatelessWidget {
-  const _GradientBox({
-    required this.item,
-  });
-
-  final NewsModel item;
-
-  @override
-  Widget build(BuildContext context) {
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topCenter,
-          end: Alignment.bottomCenter,
-          colors: [
-            ColorCommon(context).whiteAndBlackForTheme.withOpacity(0),
-            ColorCommon(context).whiteAndBlackForTheme.withOpacity(0.4),
-            ColorCommon(context).whiteAndBlackForTheme.withOpacity(0.7),
-            ColorCommon(context).whiteAndBlackForTheme.withOpacity(0.9),
-          ],
-        ),
-      ),
-      child: _TextArea(item: item),
-    );
-  }
-}
-
-class _TextArea extends StatelessWidget {
-  const _TextArea({
-    required this.item,
-  });
-
-  final NewsModel item;
-
-  @override
-  Widget build(BuildContext context) {
     return Padding(
-      padding: const PagePadding.allLow(),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          _NewsTitle(item: item),
-          const EmptyBox.xSmallHeight(),
-          _NewsDateTime(item: item),
-        ],
+      padding: const PagePadding.generalCardAll(),
+      child: SizedBox(
+        height: context.sized.dynamicHeight(0.3),
+        child: InkWell(
+          onTap: onTap,
+          child: Stack(
+            children: [
+              Positioned.fill(
+                child: _NewsImage(item: item),
+              ),
+              Positioned(
+                bottom: 0,
+                left: 0,
+                right: 0,
+                child: _TransparentBox(item: item),
+              ),
+            ],
+          ),
+        ),
       ),
     );
   }
 }
 
-class _NewsImage extends StatelessWidget {
+final class _NewsImage extends StatelessWidget {
   const _NewsImage({
     required this.item,
   });
@@ -107,16 +50,37 @@ class _NewsImage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Hero(
       tag: ValueKey(item.documentId),
-      child: CustomNetworkImage(
-        imageUrl: item.image,
-        fit: BoxFit.cover,
+      child: ClipRRect(
+        borderRadius: CustomRadius.large,
+        child: CustomNetworkImage(
+          imageUrl: item.image,
+          fit: BoxFit.cover,
+        ),
       ),
     );
   }
 }
 
-class _NewsTitle extends StatelessWidget {
-  const _NewsTitle({
+class _TransparentBox extends StatelessWidget {
+  const _TransparentBox({required this.item});
+
+  final NewsModel item;
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      shape: const RoundedRectangleBorder(
+        borderRadius: CustomRadius.large,
+      ),
+      margin: EdgeInsets.zero,
+      color: ColorCommon(context).blackAndWhiteForTheme.withOpacity(0.45),
+      child: _NewsInformationArea(item: item),
+    );
+  }
+}
+
+class _NewsInformationArea extends StatelessWidget {
+  const _NewsInformationArea({
     required this.item,
   });
 
@@ -124,37 +88,36 @@ class _NewsTitle extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    return Padding(
+      padding: const PagePadding.horizontalLowSymmetric() +
+          const PagePadding.onlyTopMedium(),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          _NewsTitle(title: item.title ?? ''),
+        ],
+      ),
+    );
+  }
+}
+
+final class _NewsTitle extends StatelessWidget {
+  const _NewsTitle({
+    required this.title,
+  });
+
+  final String title;
+
+  @override
+  Widget build(BuildContext context) {
     return Text(
-      item.title ?? '',
+      title,
       maxLines: 2,
       textAlign: TextAlign.left,
       overflow: TextOverflow.ellipsis,
-      style: context.general.textTheme.headlineSmall?.copyWith(
+      style: context.general.textTheme.titleLarge?.copyWith(
+        color: ColorCommon(context).whiteAndBlackForTheme,
         fontWeight: FontWeight.bold,
-        color: ColorCommon(context).blackAndWhiteForTheme,
-      ),
-    );
-  }
-}
-
-class _NewsDateTime extends StatelessWidget {
-  const _NewsDateTime({
-    required this.item,
-  });
-
-  final NewsModel item;
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      CustomDateTimeFormatter.formatValueTr(
-        item.createdAt ?? DateTime.now(),
-      ),
-      textAlign: TextAlign.left,
-      overflow: TextOverflow.ellipsis,
-      style: context.general.textTheme.titleSmall?.copyWith(
-        fontWeight: FontWeight.bold,
-        color: ColorCommon(context).blackAndWhiteForTheme,
       ),
     );
   }
